refactor: StreamTitleWithJammers and StreamerCategoryWithJammers
